Optimise the initial algorithm.
One edit away (Modification) : Use a flag bit to keep track of the first difference. If the flag bit has been set as True, when another difference is encountered then return
False since it's the second difference.

One insert away: The first argument is the largest string and the second argument is the smaller one. If a character difference is encountered (THE FIRST TIME), increment the second variable. If a character diff is encountered for the second time, check if indexes are equal. If they aren't equal, then that means it is the second character difference
then return False.