* Implement a rough Linked Lists
* Remove duplicate with buffer and without buffer

With buffer: Store all the values in a list and remove the node which
has the value already in the list. It gives us the performance of O(n)

Without buffer: Get value of a node and then iterate through the linked
list checking if the value exists, delete the value if it does. Do this
for all the elements of the list. Gives us the performance of O(n^2)
I didn't solve this on my own, I cheated.