kth element from the end
Two methods:

1. Recursive: Traverse recursively till we reach the end of the Linked
List. When returning back from the end, increment the return counter.
We can only print the value of the kth element from the LL and can't get
the counter.

2. Iteratively: Use two pointers to keep track. Keep the first pointer
at the kth position of the Linked List. Then traverse till the first
pointer hits the end. The second pointer which started from the beginning
would be at the kth position.