Palindrome of a string
Method 1: Adarsh's recursion.

You start from the beginning and you recurse till the end all the
while passing the start pointer till the end. Then on
your way back, you increment the start pointer and you compare that
value with the value of the recursion we're on from the end.

Method 2: Runner method

This is much easier to understand IMO. You have two pointers, a slow and
fast. The fast ptr moves twice as fast so that by the end the fast
pointer reaches the end the slow pointer would be half way through the
list. You should've pushed the values of the stack pointer to a stack

Then you compare the rest of the slow pointer till the end and pop from
the stack.