Intersection
Singly linked lists which has an intersection would have the same nodes
after the intersection. So we need to find the common nodes in the
linked lists. Get the length of both the linked lists and then take
the longer linked list and traverse the difference of the two lengths
then compare all the nodes until the intersection is found.