Since we have only two stacks, we use the new stack to push items into and then when a push action is performed, we push all the elements into the new stack and then take the top element and push rest of them back.
This is a good approach but it can be optimized. Suppose we push
all the elements of the new stack into the old stack (temporary)
and we take the oldest element from that, we need not push the
rest of the elements back to the new stack because the oldest
elements are still on the top of the old stack. Which is much
more efficient because the next time you try dequeuing an element
you just have to pop from the old stack and once it's empty,
you'll have to once more push the new stack over to the old
stack.