Brute force approach by Adarsh
So you have two stacks you can use to sort.
So basically the mStack stores all the final result.
The first iteration finds the biggest element from the stack and takes
it to the bottom. Then the stack traverses
till the second last element and finds the biggest one from there.
So one by one it brings the largest item to the botton, and finally we're
left with a sorted stack with the smallest item on top